backend/requirements.txt
fastapi
uvicorn[standard]
pandas
numpy
scipy
scikit-learn
openpyxl
python-dotenv
openai
pymoo
networkx
statsmodels
httpx
dowhy
lightgbm
tslearn
shap
ruptures
